# repos/main-meat-n-sea/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LocalMart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
    post:
      operationId: createShop
      tags: [shops]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
...
  /shops/{id}:
    get:
      operationId: getShop
      tags: [shops]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopWithProducts"
    patch:
      operationId: updateShop
      tags: [shops]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/{id}/products:
    get:
      operationId: listShopProducts
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # PRODUCTS
...
components:
  schemas:
...
    Shop:
      type: object
      required: [id, ownerId, name, slug, category, area, status]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        area:
          type: string
        logoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, suspended]
        createdAt:
          type: ["string", "null"]
    ShopWithProducts:
      allOf:
        - $ref: "#/components/schemas/Shop"
        - type: object
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
    ShopInput:
      type: object
      required: [name, slug, category, area]
      properties:
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: string
        area:
          type: string
        logoUrl:
          type: string
    ShopUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        area:
          type: string
        logoUrl:
          type: string
...
    Product:
      type: object
      required: [id, shopId, name, price, inStock]
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        inStock:
          type: boolean
        createdAt:
          type: ["string", "null"]
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        inStock:
          type: boolean
...